fix(frontend-paddle): Moves tests according to the namespace correction in 9e62a00
