+++ ivy_tests/test_ivy/test_frontends/test_paddle/test_stat.py
